파일 구조 변경(더 세세하게 나눔
